Admin: Move actions to `src/actions` folder
